backend/requirements.txt
onnxruntime
Flask
flask-cors
rembg
transparent_background
accelerate==1.2.0
aggdraw==1.3.19
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asttokens==3.0.0
attrs==24.2.0
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-help-colors==0.9.4
click-shell==2.1
colorama==0.4.6
coloredlogs==15.0.1
colormath==3.0.0
colory==0.2
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
# diffusers @ git+https://github.com/huggingface/diffusers.git@6131a93b969f87d171148bd367fd9990d5a49b6b
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imaginAIry==15.0.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.2
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
onnxruntime==1.20.1
open_clip_torch==2.29.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum-quanto==0.2.6
packaging==24.2
pandas==2.2.3
parso==0.8.4
photoshop-python-api==0.22.9
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.0
psd-tools==1.10.4
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyMatting==1.1.13
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
replicate==1.0.4
reportlab==4.2.5
requests==2.32.3
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
spandrel==0.4.0
stack-data==0.6.3
starlette==0.41.3
svglib==1.5.1
svgwrite==1.4.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.12
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchdiffeq==0.2.5
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
zipp==3.21.0